Fix sonar issues and adds/splits tests
